Split source product status tests
